backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
